improved the program structure
